Cleanup and Input test created
Cleaned up old test files, incomplete implementations of curses.
Implemented input field and created test
