+++ README.md
@@ -1,11 +1,11 @@
-# ![logo_small.png](res/logo_small.png) Discline
+# ![logo_small.png](res/logo/logo_small.png) Discline
 ------------------------------
 
-![screenshot_main.png](res/screenshot_main.png)
+![screenshot_main.png](res/screenshots/screenshot_main.png)
 
 __**Warning**__: Currently Linux/Mac only, it may be a while before support for Windows comes back
 
-**NOTE**: A total rewrite of Discline in curses/ncurses is planned. Possibly also in another language.
+**NOTE**: A total rewrite of Discline in curses/ncurses is underway!
 
 ## How to use:
 -------------------------
@@ -34,9 +34,9 @@ __**Warning**__: Currently Linux/Mac only, it may be a while before support for 
 
     If all else fails, join the dev discord and we'll be glad to help!
 
-4. Run `--store-token` to store your token
+4. Run `python3 Discline.py --store-token` to store your token
 
-5. Run `--copy-skeleton` to get a template config
+5. Run `python3 Discline.py --copy-skeleton` to get a template config
 
 6. Edit `~/.config/Discline/config` to your choosing.
 
@@ -108,6 +108,19 @@ __**Warning**__: Currently Linux/Mac only, it may be a while before support for 
 Almost all aspects of the client can be colored to
 the user's wishes. You can set these colors from within `~/.config/Discline/config`
 
+Note: These assume that you're using the standard terminal colors. If you
+have colors already defined in your ~/.Xresources or similar, this will 
+be very confusing.
+
+## Launching
+------------------------
+Discline uses git for automatic updates, so you must be within the Discline
+directory upon starting. Manually you can launch via `python3.6 ./Discline.py`, 
+however it is advised to create a helper script to do this for you.
+
+An example script is in the /res/scripts folder, 
+edit it to suit your system and tastes.
+
 ### A Note On Emojis
 -------------------------
 
@@ -129,7 +142,7 @@ Custom emojis however, are displayed as :emoji_name:
 Like emojis, not all terminals and fonts support
 italic/bold/underline and 'background' colors, (which are used for \`code\`).
 If these features aren't working for you, odds are you are not using a
-support terminal/font. Experiment with different setups to see what works.
+supported terminal/font. Experiment with different setups to see what works.
 
 ![Image](https://0x0.st/sHQ0.png)
 
@@ -199,7 +212,8 @@ and make sure it gets sourced upon opening your terminal.
 ### Misc Screenshots
 --------------------------
 
-![Image](res/screenshot2.png)
+![Image](res/screenshots/kingk22-screenshot.png)
+
 ![Image](https://0x0.st/sH5g.png)
 
 ![Image](https://0x0.st/sHjn.png)
@@ -208,9 +222,9 @@ and make sure it gets sourced upon opening your terminal.
 
 ![Image](https://0x0.st/sHez.png)
 
-It can even be configured to hide elements of the UI in the config:
+It can even be configured to hide most elements of the UI in the config:
 
-![Image](https://i.imgur.com/tFv9Dpc.png)
+![Image](res/screenshots/minimal_brown_ss.png)
 
 ### Known Bugs
 --------------------------
