--- curses_tests/escTester.py
@@ -1,1 +1,0 @@
-print("\033[1mTest\033[0mTest")
