--- curses_tests/lorem.json
@@ -1,17 +1,0 @@
-[
-	{"name":"Alex",
-		"content":"Test content abcde"
-	},
-	{"name":"John",
-		"content":"Lorem ipsum dolor sit amet, consectetur adipiscing elit."
-	},
-	{"name":"Felix",
-		"content":"In sit amet tincidunt nisi. Duis eros ipsum, porttitor gravida eros ut, viverra egestas nisi. Nam pulvinar dui odio, non ornare lacus lobortis et."
-	},
-	{"name":"Katie",
-		"content":"Aenean sollicitudin pellentesque cursus. Donec id justo ut orci efficitur imperdiet."
-	},
-	{"name":"Nat",
-		"content":"Suspendisse nec dapibus erat, vestibulum sollicitudin dolor. In sollicitudin tortor at scelerisque ullamcorper."
-	}
-]
