--- curses_tests/lorem.py
@@ -1,3 +1,0 @@
-from loremipsum import get_sentence
-
-print(type(get_sentence()))
