--- curses_tests/poptest.py
@@ -1,9 +1,0 @@
-a = ["Oldest", "1", "2", "3", "4", "Newest"]
-
-#Existing msgs
-# "Oldest" (0)
-# ...
-# "Newest" (5)
-
-while a:
-    print(a.pop(0))
